+++ tests/client/test_commands.py
@@ -1810,6 +1810,10 @@ class TestRedisCommands(object):
         assert await r.geohash('barcelona', 'place1', 'place2') ==\
             [b'sp3e9yg3kd0', b'sp3e9cbc3t0']
 
+    @skip_if_server_version_lt('4.0.0')
+    def test_geopos_no_value(self, r):
+        assert r.geopos('barcelona', 'place1', 'place2') == [None, None]
+
     @skip_if_server_version_lt('3.2.0')
     @pytest.mark.asyncio(forbid_global_loop=True)
     async def test_geopos(self, r):